Revert Teleport + formatter run
