+++ CS2Fixes.vcxproj
@@ -221,7 +221,7 @@
     <ClCompile Include="src\utils\weapon.cpp" />
     <ClCompile Include="src\utils\hud_manager.cpp" />
     <ClCompile Include="src\utils\utils.cpp" />
-    <ClCompile Include="src\utils\script_function.cpp" />
+    <ClCompile Include="src\utils\vscript_function.cpp" />
     <ClCompile Include="src\cs2_sdk\entity\services.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -295,7 +295,7 @@
     <ClInclude Include="src\utils\entity.h" />
     <ClInclude Include="src\utils\module.h" />
     <ClInclude Include="src\utils\plat.h" />
-    <ClInclude Include="src\utils\script_function.h" />
+    <ClInclude Include="src\utils\vscript_function.h" />
     <ClInclude Include="src\utils\virtual.h" />
     <ClInclude Include="src\cs2_sdk\netmessages.h" />
     <ClInclude Include="src\utils\weapon.h" />
