+++ gamedata/cs2fixes.jsonc
@@ -397,6 +397,11 @@
 			"windows": 150,
 			"linux": 149
 		},
+		"Teleport":
+		{
+			"windows": 163,
+			"linux": 162
+		},
 		"CollisionRulesChanged":
 		{
 			"windows": 186,
